backend/ocr/requirements.txt
# AI ID OCR Service for HIS
# Lightweight dependencies - No torch/transformers

# FastAPI framework
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6

# Google Gemini for OCR extraction
google-generativeai>=0.3.0

# Image processing
pillow>=10.1.0
pytesseract>=0.3.10

# Utilities
python-dotenv>=1.0.0
httpx>=0.25.0
